fix: extract shared Jest boilerplate into setupFilesAfterEnv to remove duplication
