+++ jest.config.cjs
@@ -6,4 +6,5 @@ module.exports = {
   moduleNameMapper: {
     '^@/(.*)$': '<rootDir>/$1',
   },
+  setupFilesAfterEnv: ['<rootDir>/tests/setup.ts'],
 };
